pyproject.toml
[tool.poetry]
name = "network-sec-project"
version = "0.1.0"
description = ""
authors = ["Jeddy Tan <tanjd@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
pylint = "^2.17.5"
isort = "^5.12.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.black]
line-length = 120
target-version = ['py38']

# pytest.ini
[tool.pytest.ini_options]

[tool.pylint.master]
# fix relative import
init-hook = 'import sys; sys.path.append("./")'

[tool.pylint.messages_control]
# ignoring (C)onvention, (R)efactor,
# (W0511) todo/fixme
# (W1203) % lazy formatting in logging
# (W0105) multiline comment
disable = "R,C,W0511,W1203,W0603,W0105,W0621,W0622,E0213"
extension-pkg-whitelist = "pydantic"
extension-pkg-allow-list = "ujson"

[tool.pylint.format]
max-line-length = "120"
